double instead float

+++ AndroidBLE/app/src/main/java/com/pascaldornfeld/gsdble/overview/components/CharaDataStatsByGradient.kt
@@ -6,11 +6,11 @@ import kotlin.math.sqrt
 
 class CharaDataStatsByGradient(
     timestamp: Long,
-    averageGraph: GraphFragment<Float>?,
-    deviationGraph: GraphFragment<Float>?
+    averageGraph: GraphFragment<Double>?,
+    deviationGraph: GraphFragment<Double>?
 ) : CharaDataStatsCalculatorAsyncTask(timestamp, averageGraph, deviationGraph) {
 
-    override fun doInBackground(vararg params: Array<Pair<Long, Long>>?): Float? {
+    override fun doInBackground(vararg params: Array<Pair<Long, Long>>?): Double? {
         assert(params.isNotEmpty())
         val inputArray = params[0]
 
@@ -24,23 +24,23 @@ class CharaDataStatsByGradient(
                     (pair.second.toDouble() - prev.second.toDouble()) / (pair.first.toDouble() - prev.first.toDouble())
                 }
             }.filterNotNull().average()
-            publishProgress(avgTime.toFloat())
+            publishProgress(avgTime)
 
             // calculate variance
-            var sum = 0.0f
+            var sum = 0.0
             inputArray.forEachIndexed { index, pair ->
                 if (index != 0) {
                     val prev = inputArray[index - 1]
                     val gradient =
                         (pair.second.toDouble() - prev.second.toDouble()) / (pair.first.toDouble() - prev.first.toDouble())
-                    sum += (gradient - avgTime).pow(2).toFloat()
+                    sum += (gradient - avgTime).pow(2)
                 }
                 if (isCancelled) return null
             }
 
             // return deviation
             // sum is based on inputArray.size -1 elements. The formula takes this amount -1. so we got inputArray.size -2
-            return sqrt((sum / (inputArray.size - 2).toFloat()))
+            return sqrt((sum / (inputArray.size - 2)))
         } else {
             publishProgress(null)
             return null
